Matlab Fixing for single timestep
I fixed an index issue that was from the input to Signal_Cascade. I'm
not sure everything I did is correct though. I went on the assumption
that all of the variables in the signal cascade were shared between the
microvilli. If that is indeed the case, the code might actually be
working. The final I_in found was 0.68, which is pretty close to the
input we'd been using for HH (0.6). Don't know the units though, mA?