It's ALIIIIIIVE! ....in Matlab.
Main thing was making all of the various components work together.
Specify length of run at first for loop variable runtime, n_m sets how
many microvilli to run with. 30000 might be too many, especially for
those of you with laptops =).

Made RPAM a function. For some reason, when it was just a script it
would clobber the initialization of the X vector. I also think it caused
a memory leak big time, I'm talking 7+ GB.